Correct area- and volume-averages for partially-wet cases
The multiplication by the wet mask in this line:
```python
A = grid.get_metric(da, ["X", "Y"])*ds.wet
```
was incorrect because it meant that the weights used for
area/volume-weighted averages were NaN (or zero) over land,
and thus represented ocean-area or ocean-volume-averages, which is
not what one would expect from this coarsening operation. While this
could in principle be accounted for later by re-multiplying by the
wet area of each cell, this incorrect averaging also affected the
coarsened wet mask, meaning that it was 1 for completely wet cells
and NaN for any coarse cell that had ocean in any of the its contributing
native-resolution cells.

I fixed this by removed the multiplication by the wet mask when computing the
weights for the averaging.

Some other small bug-fixes:
- I now add some calls to the `fillna(0.)` method to ensure that NaNs do not
propagate.
- When I override the coarsened datasets coords, I now create the xr.DataArray
manually because I was having problems where the different coordinates would override
eachother. In particular, `areacello` was incorrectly being overwritten with the
average of all contributing cells rather than their sum.
- Added a provenance comment that cites the version of `CM4Xutils` used for the data
processing.
- Overwrite the misleading `nominal longitude` and `nominal latitude` coordinates with
integer indices (this is consistent with new OM5 conventions).