Conservatively transform to sigma2 coordinates and coarsen horizontally (#3)
* Function for regridding any variables in any vertical coordinate to `sigma2`

Handles both intensive and extensive variables.

* Functionality for conservatively coarsening a dataset in the horizontal dimensions

Uses the `cell_methods` attribute to flexibly and automatically deal with four cases:
- fully intensive (3D mean) quantities, e.g. `thetao`
- verticaly extensive but horizontally intensive quantities (2D mean), e.g. `thkcello`
- horizontally extensive quantities, e.g. `areacello`
- transports, which are snapshots in one dimension but extensive in the others

The `subsample_geocoords` utility function overwrites the geocoordinate coordinate values
with the correct subsampling. This is not trivial because it depends on whether the
coarsening amount is even or odd!

* Vertical coordinate remapping and conservative horizontal coarsening

New functions and scripts for conservatively coarsening diagnostics
(budgets and tracers), used for reducing the data volume from ~300 TB
to a more tractable ~2-3 TB for transfering to @anthony-meza.

* Correct area- and volume-averages for partially-wet cases

The multiplication by the wet mask in this line:
```python
A = grid.get_metric(da, ["X", "Y"])*ds.wet
```
was incorrect because it meant that the weights used for
area/volume-weighted averages were NaN (or zero) over land,
and thus represented ocean-area or ocean-volume-averages, which is
not what one would expect from this coarsening operation. While this
could in principle be accounted for later by re-multiplying by the
wet area of each cell, this incorrect averaging also affected the
coarsened wet mask, meaning that it was 1 for completely wet cells
and NaN for any coarse cell that had ocean in any of the its contributing
native-resolution cells.

I fixed this by removed the multiplication by the wet mask when computing the
weights for the averaging.

Some other small bug-fixes:
- I now add some calls to the `fillna(0.)` method to ensure that NaNs do not
propagate.
- When I override the coarsened datasets coords, I now create the xr.DataArray
manually because I was having problems where the different coordinates would override
eachother. In particular, `areacello` was incorrectly being overwritten with the
average of all contributing cells rather than their sum.
- Added a provenance comment that cites the version of `CM4Xutils` used for the data
processing.
- Overwrite the misleading `nominal longitude` and `nominal latitude` coordinates with
integer indices (this is consistent with new OM5 conventions).

* Bump version and add functions to coarsen tracers

- Bump version to v0.1.0
- Bundle functions for remapping and coarsening CM4X output into
new `remap_functions` script to simplify run scripts
`coarsen_sigma2_budgets.py` and `coarsen_sigma2_tracers.py`.
- Change logic for loading tracers to deal with the fact
that for some variables they come in `5yr` chunks while
others they come in `10yr` chunks... annoying.