Consistent coarsening of all variables and coordinates
Major changes:

1. Consistent coarsening of variables
- When variables with cell methods {"xh":"mean", "yh":"mean"}, such as budget diagnostics
and tracer concencentrations, are coarsened, they now accurately reflect the area- or
volume-weighted averages of the underlying *wet* cells only. Previously, we computed
averages over the area of all cells that contribute to the coarsened cell, including
land cells. This was done so that `areacello` on the coarsened grid was just a simple
sum over all of the contributing cells and was easier to interpret. However, it made
interpreting many of the coarsened diagnostics very confusing. For example, the
coarsened `tos` (temperature of the ocean surface) was misleadingly low because it had
effectively been averaged together with zero ocean heat content over land cells.

- Instead, these area-averaged coarsened variables are indeed only averaged over ocean
cells, which means that to reconstruct tracer content or tracer content tendencies,
we need to multiplty by the coarsened *ocean-only area*. We have modified the
calculation of `areacello` on the coarsened grid to reflect this.

2. Coarsening of `wet_u`, `wet_v`, `dxCv`, and `dyCu`

3. Inclusion of ice diags
We now include some ice diags in the `load_wmt_grid` function. This was more complicated
than anticipated because the units, coordinates, and resolution of the ice diagnostics
differs from that of the ocean diagnostics... But it works for some variables now!

4. Calculation of the 3D "boundary_forcing_h_tendency" mass tendency diagnostic
from the 2D surface field "wfo" if it is missing, as for odiv-209 and odiv-210.
I had previously not realized that this diagnostic is missing for these runs and
my water mass transformation calculations for the piControl spinups of CM4X were
totally off as a result!