Fixes two bugs related to NaNs propagating into transformations
1) `xgcm.transform` does not correctly treat internal boundaries,
and seems to set any cell bounded by a NaN on one of the two interfaces
of the target data as a NaN itself. For transforming from z coordinates to
density coordinates, I think this means that all of my bottom cells are
essentially being set to NaN. I discovered this bug because I realized
that my coarsened transports were inconsistent with my coarsened transport
divergences. I fixed this bug by creating a new `fillna_below` function
that fills the value of all cells below the bottom grid cell with the
value of the bottom grid cell. I confirmed that this preserves vertical
integrals.

2) `xgcm.interp` also does not correctly treat internal boundaries (e.g.
between land and ocean). I created an interpolation function that just
averages the tracer values of the cells on either side of valid velocities.