Fix bug in `fillna_below` function
Accidentally removed the vertical coordinate argument of the
`da.argmax` method call that helps us find the index of the
bottom-most valid cells! This was using a single global index
which has the result of filling cells below the bottom-most one
with the surface-most cell value, rather than the bottom-most!
It gave very incorrect results!