Fix fillna_below function
This fixes the fillna_below function and improves its performance
by assuming the 3D wet mask is constant in time (i.e. it reads the
3D (lon, lat, vertical coordinate) mask from the first time index).

This is meant to be able to work regardless of the actual names of the
coordinates, although it assumes that vertical coordinate is monotonic with
depth and that k=0 at the surface.

This change improves the comparison between the dora and zarr WMT calculations,
although it is still not exact, I think because the dora calculation is
actually slightly flawed because this fillna_below function is not used
in `xwmb` and so some cells in the original z-coordinate dataset are
being effectively turned to NaNs when we try to transform to vertical coordinates.
This should now be fixed in the transformation functions in CM4Xutils, which
means that the density-binned diagnostics it produces should be correct.