Correct treatment of boundary conditions in transformation to density coordinates (#7)
* Fixes two bugs related to NaNs propagating into transformations

1) `xgcm.transform` does not correctly treat internal boundaries,
and seems to set any cell bounded by a NaN on one of the two interfaces
of the target data as a NaN itself. For transforming from z coordinates to
density coordinates, I think this means that all of my bottom cells are
essentially being set to NaN. I discovered this bug because I realized
that my coarsened transports were inconsistent with my coarsened transport
divergences. I fixed this bug by creating a new `fillna_below` function
that fills the value of all cells below the bottom grid cell with the
value of the bottom grid cell. I confirmed that this preserves vertical
integrals.

2) `xgcm.interp` also does not correctly treat internal boundaries (e.g.
between land and ocean). I created an interpolation function that just
averages the tracer values of the cells on either side of valid velocities.

* Bump minor version to reflect changes to the data processing

* Apply fillna_below bugfix to non-velocity fields too!

* Fix bug in `fillna_below` function

Accidentally removed the vertical coordinate argument of the
`da.argmax` method call that helps us find the index of the
bottom-most valid cells! This was using a single global index
which has the result of filling cells below the bottom-most one
with the surface-most cell value, rather than the bottom-most!
It gave very incorrect results!

* Fix fillna_below function

This fixes the fillna_below function and improves its performance
by assuming the 3D wet mask is constant in time (i.e. it reads the
3D (lon, lat, vertical coordinate) mask from the first time index).

This is meant to be able to work regardless of the actual names of the
coordinates, although it assumes that vertical coordinate is monotonic with
depth and that k=0 at the surface.

This change improves the comparison between the dora and zarr WMT calculations,
although it is still not exact, I think because the dora calculation is
actually slightly flawed because this fillna_below function is not used
in `xwmb` and so some cells in the original z-coordinate dataset are
being effectively turned to NaNs when we try to transform to vertical coordinates.
This should now be fixed in the transformation functions in CM4Xutils, which
means that the density-binned diagnostics it produces should be correct.

* Scripts for transfering to collab for Globus transfers

* Updated example notebook

* Copy files to collab