Extend target sigma2 bins to ensure mass conservation!
Previously, `sigma2_i` bins used in the vertical regridding
only went from -3 kg/m^3 to about 40 kg/m^3. While this
range includes almost all ocean densities, it excludes very salty
and hence very dense waters in marginal seas, such as the Red Sea.

The resulting errors in mass conservation are visible in the figure in
the `Confirmation that ocean volume and layer thickness are preserved`
section of this notebook:
https://github.com/hdrake/CM4Xutils/blob/ab6f442c6b99856cdbd2faa75f1f891e21a27d0e/examples/coarsen_budget_diags.ipynb

This PR fixes this issue by adding extremely generous bins to both the
lower and upper bounds of the target sigma2 grid.