Fix area-mask bug in coarsening of z-coord MMS diags
This fixes a bug that only affected the coarsening of variables with
cell method mean-mean-sum (horizontal averages but vertical sums), such
as thkcello and vertically-integrated tracer tendencies, in depth
coordinates. An incorrect 3D area mask was being used. This bug does not
affect diagnostics that have already been mapped into density coordinates
before coarsening because in those, nans have already been filled with zeros,
so the mask is all ones.